"core" bashrc.d functions for controlled loading
the bashrc.d control structure is simply a for loop within a .bashrc
file. By moving the "load" logic into its own function, we can open the
door for more control in the future. Most notably:
	- loaded components are remembered, so we can load-on-demand
	- we have defined a naming convention for component functions
	- most importantly, there is now a central point for adding more
	  complex logic to the "load" function

None of this code is "live" as of this commit. This is just laying the
groundwork.