remove trailing slash from .grepignore exclude-dir paths
when specifying entries within a .grepignore file, a trailing
slash is used to denote that the entry is a "directory", rather
than a "file" pattern. As this syntax is specific to .grepignore,
it should not be passed along to grep with the trailing slash
preserved. This commit ensures that the trailing slash is removed
from entries passed into --exclude-dir