Allow for per-project GOPATH specification
now .goworkspace is used as an indicator for the root of a go project.
GOPATH will default to the nearest location of .goworkspace, or the
path(s) pointed to by the .goworkspace file. $HOME/.goworkspace may
contain "global" paths to include by default.

Within a .goworkspace file, paths can be specified relative to the
.goworkspace file by starting paths with a "./". This will be expanded
to the absolute path prior to inclusion in the GOPATH.

If the .goworkspace indicator is empty, the paths within
$HOME/.goworkspace are not appended to the generated GOPATH. If you want
the indicator-derived GOPATH + the $HOME/.goworkspace paths, fill the
.goworkspace indicator with a single period (".")

This commit also adds the generic "uproot" utility, which is used to
find the indicator file.

$ uproot indicator-filename

will recursively search upwards until a file named "indicator-filename"
is located, and will output an appropriate number of "../.." to reach
the filename, such that "$(uproot somefile)/somefile" will always refer
to "somefile" itself (if it exists at all)